refactor(tv): extract shared LinkModal from QuizButtonRenderer
Extract QrMatrix, AndroidTvWebViewContent, TvOSQrContent, and modal
wrapper into a reusable LinkModal component. QuizButtonRenderer now
imports LinkModal and passes quiz-specific props (isAllowedQuizUrl
validator, error text). All WebView security hardening preserved.
